renamed a file variable
